// src/modules/products/types.ts
export namespace IEntity {
  export interface IProduct {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: Array<string>;
    brand: string;
    weight: number;
    dimensions: Dimensions;
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: Array<Review>;
    returnPolicy: string;
    minimumOrderQuantity: number;
    meta: Meta;
    images: Array<string>;
    thumbnail: string;
  }

  export interface Dimensions {
    width: number;
    height: number;
    depth: number;
  }

  export interface Review {
    rating: number;
    comment: string;
    date: string;
    reviewerName: string;
    reviewerEmail: string;
  }

  export interface Meta {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
  }
}